+++ bin/const.ts
@@ -1,2 +1,2 @@
-export const RN_VERSION = '0.62.0';
+export const RN_VERSION = '0.62.2';
 export const EXPO_VERSION = '37';